+++ src/linklayer/can/CanPortInput.h
@@ -80,9 +80,9 @@ private:
     vector<int> outgoingRemoteFrameIDs;
 
     /**
-     * @brief Valid values are between 10000 and 1000000. Initialized from ned-attribute of CAN-Bus.
+     * @brief Bandwidth of the bus in Mbps.
      */
-    int bandwidth;
+    double bandwidth;
 
     /**
      * true if errors are activated. Initialized from ned-attribute of CAN-Node.
